Refactor function compilation into modular, specialized function modules
